HOMEWORK-07: Рефакторинг кода + добавила комментарии
